even more Harl

+++ ex05/main.cpp
@@ -4,11 +4,11 @@
 int main(void)
 {
     Harl harl;
-    int complain_amount = std::numeric_limits<int>::max() - 1;
-    for (int i = std::numeric_limits<int>::min() + 1; i < complain_amount; i++)
+    int64_t complain_amount = std::numeric_limits<int64_t>::max() - 1;
+    for (int64_t i = std::numeric_limits<int64_t>::min() + 1; i < complain_amount; i++)
     {
         std::cout << "i: " << i << "/" << complain_amount << ": ";
-        int complain_level = std::abs(i % 4);
+        int64_t complain_level = std::abs(i % 4);
         switch (complain_level)
         {
             case 0: harl.complain("DEBUG"); break;
